src/main/java/lapr/project/model/service/CourierService.java
package lapr.project.model.service;

import lapr.project.data.CourierDB;
import lapr.project.model.Courier;
import lapr.project.model.PassGenerator;
import lapr.project.model.Pharmacy;

import java.security.NoSuchAlgorithmException;

public class CourierService {

    private CourierDB oCourierDB;

    public CourierService() {
        oCourierDB = new CourierDB();
    }

    public Courier newCourier(String strName, String strEmail, Integer strNIF, String strIBAN, Pharmacy oPharmacy) throws NoSuchAlgorithmException {
        PassGenerator pass = new PassGenerator();
        String password = pass.generatePassword();

        return new Courier(strName, strEmail, password, strNIF, strIBAN,oPharmacy);
    }

    public boolean registersCourier(Courier oCourier) {
        return oCourierDB.addCourierToDB(oCourier.getName(), oCourier.getEmail(), oCourier.getPw(), oCourier.getNif(), oCourier.getM_iban(), oCourier.getM_Pharmacy().getId());
    }
}
